Optional configuration for specific project_paths eval.
GitLab server 13.2 added a GL_PROJECT_PATH environment variable that
conveys the path for the target project. This addition introduces an
optional configuration for “evaluate_prefixes_only" and with that a
"project_paths" list. Defining this will check the established path
against a list of proposed prefixes and only by containing one such
prefix will the mirror evaluation be observed.
The goal is to allow instances with already established
project/mirroring policies to experiment with remote mirror security on
a more targeted set of groups.