Return an empty list of collabs when unauthorized
When the hook is repeatedly called on a repository where it does not
have permissions to access the organization, it will consistently get
unauthorized HTTP responses back. It won't change over time and at worst
it could cause the webserver to limit requests. By returning an empty
list, this change allows sets the cache to be empty and thus not make
repeated requests that will fail.